upstream release 10.17 (pre-built binary)
This packs a new 10.17 release, but it no longer builds from source, it
includes the upstream-provided pre-built binary instead. There are many
pain-points with this upstream, but the biggest issue is that the author
modifies system libraries (like gtk, wxwidgets, curl, libssh) and
expects everyone to modify their libraries as well (or use the pre-built
binary, of course). These patches aren't even in a usual diff format,
they are all documented in a free-form text document (see Bugs.txt in
the source archive). These patches aren't pending bug-fixes, they are
often RFEs which have been rejected by the particular project several
years ago. This is a clear no-go in how our open-source distribution
works today. Without these patches, the source code often fails to
compile. It is necessary to revert such modifications, which often
requires to replace the code with 'same in functionality, but different'
code. This makes it very maintenance-heavy and requires deep
understanding of the project and its libraries.

After trying to patch and compile version 10.16 and then 10.17 from
source, I decided the only option is to go with binary releases starting
from 10.17, or there needs to another volunteer to fight with the broken
upstream practices and the code, if anyone cares enough to have this
compiled from scratch.