use extra-data step to download the FFS binary (second try)
This is another attempt at using extra-data, first tried in c177e81541 and then
reverted in b8c79651426 (see both descriptions). Turns out a system-wide
installation runs under fake root, which tries to restore original UID for files
in a tar archive and fails to do that. The solution was to append
`--no-same-owner` to the `tar` command.

Fixes: https://github.com/flathub/org.freefilesync.FreeFileSync/issues/48
Fixes: https://github.com/flathub/org.freefilesync.FreeFileSync/issues/53