allow access to ~/.var/app
This is needed for people wanting to back up the data of their Flatpak apps.
It seems the access would be possible through file-open dialog portal, if FFS
used a recent enough GUI toolkit. But for the moment, it needs to be hardcoded.

Fixes https://github.com/flathub/org.freefilesync.FreeFileSync/issues/47